sessions/7/egor_fedorov/requirements.txt
flask-socketio
flask
pytest
websocket-client
eventlet